% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthlyInt}
\alias{monthlyInt}
\title{Monthly Intervening Natural Flow}
\format{An xts matrix with 1287 rows (months) and 29 columns (locations).
All values are in acre-ft.}
\source{
\url{http://www.usbr.gov/lc/region/g4000/NaturalFlow/current.html}
}
\usage{
monthlyInt
}
\description{
A data set containing the monthly intervening natural flow at 29 locations 
in the Colorado River Basin from October 1905 through December 2012.
}
\details{
Natural flow represents the flow that would have  occurred at a location, had 
depletions and reservoir regulation not been present upstream of that location. 
The total natural flow represents the total flow in the entire basin above a given 
gage. The intervening natural flow represents only the flow that originates 
upstream of a given gage, but below all upstream gage locations.

The natural flows are recomputed annually and are subject
to change without notice. All methods are fully desribed at 
\url{http://www.usbr.gov/lc/region/g4000/NaturalFlow/current.html}.

Each of the 29 locations are represented by one variable (column). The variable
names use a shorthand name of each gage location, e.g., \code{LeesFerry} is
shorthand for the "Colorado River at Lees Ferry, AZ". See \code{CRSSIO::nfGageNames()}
for full gage names, and \code{CRSSIO::nfShortNames()} for the shorthand version used
to name the variables in these data.

The data includes xts attributes (\code{xtsAttributes(monthlyTot)}) for the
source, the source workbook, and the sheet name the data were created from.
}
\seealso{
\code{\link{monthlyTot}}, \code{\link{cyAnnTot}}, \code{\link{cyAnnInt}},
\code{\link{wyAnnTot}}, \code{\link{wyAnnInt}}
}
\keyword{datasets}

