% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthlyTot}
\alias{monthlyTot}
\alias{monthlyInt}
\alias{cyAnnTot}
\alias{cyAnnInt}
\alias{wyAnnTot}
\alias{wyAnnInt}
\title{Monthly and Annual; Total and Intervening Natural Flow}
\format{An xts matrix with 1323 (monthly data) or 110 (annual data) rows
and 29 columns (locations). All values are in acre-ft. For annual data, each
value is associated with December of the year for calendar year data and
September of the year for water year data.}
\source{
\url{http://www.usbr.gov/lc/region/g4000/NaturalFlow/current.html}
}
\usage{
monthlyTot

monthlyInt

cyAnnTot

cyAnnInt

wyAnnTot

wyAnnInt
}
\description{
A data set containing the monthly and annual; total and intervening natural
flow at 29 locations in the Colorado River Basin from October 1905 through
December 2016.
}
\details{
Natural flow represents the flow that would have  occurred at a location, had
depletions and reservoir regulation not been present upstream of that
location. The total natural flow represents the total flow in the entire
basin above a given gage. The intervening natural flow represents only the
flow that originates upstream of a given gage, but below all upstream gage
locations.

The natural flows are recomputed annually and are subject to change without
notice. All methods are fully desribed at
\url{http://www.usbr.gov/lc/region/g4000/NaturalFlow/current.html}.

Each of the 29 locations are represented by one variable (column). The
variable names use a shorthand name of each gage location, e.g.,
\code{"LeesFerry"} is shorthand for the "Colorado River at Lees Ferry, AZ". See
\code{nf_gage_names()}  and \code{nf_gage_abbrv()} in the CRSSIO package for
full gage names and shorthand versions, respectively, used to name the
variables in these data.

The monthly data will sum to equal the annual data for each location as
follows:
\itemize{
\item Summing \code{monthlyTot} each year from January through December
will equal \code{cyAnnTot}
\item Summing \code{monthlyInt} each year from January through December
will equal \code{cyAnnInt}
\item Summing \code{monthlyTot} each water year from October through
September of the following year will equal \code{wyAnnTot}
\item Summing \code{monthlyInt} each water year from October through
September of the following year will equal \code{wyAnnInt}
}

The data include xts attributes (ex: \code{xtsAttributes(monthlyTot)}) for the
source, the source workbook, and the sheet name the data were created from.
}
\keyword{datasets}
